+++ src/pdi/operations/areaWise/AreaWiseOperation.java
@@ -3,5 +3,5 @@ package pdi.operations.areaWise;
 import java.util.List;
 
 public interface AreaWiseOperation {
-    short apply(List<Short> values, short maxValue);
+    int apply(List<Integer> values, int maxValue);
 }
