+++ src/pdi/operations/elementWise/ElementWiseOperation.java
@@ -1,5 +1,5 @@
 package pdi.operations.elementWise;
 
 public interface ElementWiseOperation {
-    short apply(short value, short maxValue);
+    int apply(int value, int maxValue);
 }
